requirements.txt
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==0.21.1
Werkzeug==2.2.3
apispec==6.0.2
dominate==2.7.0
Flask-Bootstrap==3.3.7.1
Flask-Moment==1.0.5
flask-smorest==0.40.0
Flask-WTF==1.1.1
gunicorn==20.1.0
marshmallow==3.19.0
packaging==23.0
visitor==0.1.3
webargs==8.2.0
WTForms==3.0.1
